DRY display of listings and sorted some functions into utils folder
