+++ src/js/components/listings/outputListing.js
@@ -1,48 +1,25 @@
-import { updateCountdown } from "../../components/countdown.js";
-
-function createImage(media, altText, defaultUrl) {
-  const img = document.createElement("img");
-  img.src = media && media.length > 0 ? media[0].url : defaultUrl;
-  img.alt = altText;
-  img.classList.add("listing-image");
-  return img;
-}
-
-function calculateCurrentBid(bids) {
-  return bids && bids.length > 0
-    ? Math.max(...bids.map((bid) => bid.amount))
-    : 0;
-}
-
-export function getListingDetails(listing) {
-  const hasExpired = new Date(listing.endsAt) < new Date();
-
-  // Get the last bid if the auction has ended, or the current bid if still active
-  const lastBid =
-    hasExpired && listing.bids && listing.bids.length > 0
-      ? listing.bids[listing.bids.length - 1].amount
-      : calculateCurrentBid(listing.bids);
-
-  return { hasExpired, lastBid };
-}
+import { updateCountdown } from "./utils/countDown.js";
+import { bidTimeDetails } from "./utils/timeBid.js";
 
 export function outputListings(listing) {
-  const { hasExpired, lastBid } = getListingDetails(listing);
+  const { hasExpired, lastBid } = bidTimeDetails(listing);
 
   // Create main container
   const listingContent = document.createElement("div");
   listingContent.classList.add("listing-content");
   if (hasExpired) listingContent.classList.add("expired");
 
-  // Add listing image
-  const mediaImg = createImage(
-    listing.media,
-    listing.title,
-    "../../src/images/logo.jpg",
-  );
+  // Listing image
+  const mediaImg = document.createElement("img");
+  mediaImg.src =
+    listing.media && listing.media.length > 0
+      ? listing.media[0].url
+      : "../../src/images/logo.jpg";
+  mediaImg.alt = listing.title;
+  mediaImg.classList.add("listing-image");
   listingContent.appendChild(mediaImg);
 
-  // Add seller container
+  // Seller container
   const sellerContainer = document.createElement("div");
   sellerContainer.id = "seller-container";
 
@@ -61,12 +38,12 @@ export function outputListings(listing) {
 
   listingContent.appendChild(sellerContainer);
 
-  // Add title
+  // Title
   const title = document.createElement("h2");
   title.textContent = listing.title;
   listingContent.appendChild(title);
 
-  // Add bid info
+  // Bid info
   const currentBid = document.createElement("p");
   currentBid.textContent = `${hasExpired ? "Winning Bid" : "Current Bid"}: ${lastBid} credits`;
   listingContent.appendChild(currentBid);
