+++ src/js/components/listings/utils/calculateBid.js
@@ -1,0 +1,5 @@
+export function calculateCurrentBid(bids) {
+  return bids && bids.length > 0
+    ? Math.max(...bids.map((bid) => bid.amount))
+    : 0;
+}
