+++ src/js/components/profile/avatar.js
@@ -1,6 +1,6 @@
 import { updateProfile } from "../../api/profile/update.js";
 import { showErrorAlert, showSuccessAlert } from "../../global/alert.js";
-import { validateImageUrl } from "../../global/validImg.js";
+import { validateImageUrl } from "../form/utils/validImg.js";
 
 export function avatarUpdate() {
   const avatarImg = document.getElementById("profile-avatar");
