+++ src/js/components/search.js
@@ -1,6 +1,6 @@
-import { displayListings } from "../api/listing/displayListings.js";
 import { searchListings } from "../api/listing/listingService.js";
 import { showErrorAlert } from "../global/alert.js";
+import { displayListings } from "./listings/displayListings.js";
 
 /**
  * Initializes the search functionality for a page.
