+++ src/js/views/pages/profile.js
@@ -1,5 +1,5 @@
-import { displayUserListings } from "../../api/listing/displayListing.js";
 import { fetchProfile } from "../../api/profile/fetchProfile.js";
+import { displayUserListings } from "../../components/listings/displayUserListings.js";
 import { avatarUpdate } from "../../components/profile/avatar.js";
 import { displayProfile } from "../../components/profile/displayProfile.js";
 import { requireAuth } from "../../global/authGuard.js";
