+++ src/js/views/pages/register.js
@@ -1,5 +1,5 @@
 import FormHandler from "../../components/form/formHandler.js";
-import { setupPreviewInputs } from "../../components/previewHandler.js";
+import { setupPreviewInputs } from "../../components/form/utils/previewHandler.js";
 
 FormHandler.initialize("#register-form", "register");
 setupPreviewInputs(
